Break out the code that handles re/destructuring for inlining in classic.  this also makes it easy to see where the seam is between the original code and the relatively complex handling of captures and destructuring for inlining
