+++ src/classic/clvm_tools/stages/stage_2/mod.rs
@@ -1,6 +1,7 @@
 pub mod compile;
 pub mod defaults;
 pub mod helpers;
+pub mod inline;
 pub mod module;
 pub mod operators;
 pub mod optimize;
